requirements.txt
python-telegram-bot==20.8
Flask==2.3.3
gspread==5.12.4
oauth2client==4.1.3
requests==2.31.0
